pyproject.toml
[tool.poetry]
name = "typed-graphql"
version = "0.22.0"
description = "Use types to have a simpler graphql implementation"
authors = ["Willem Thiart <himself@willemthiart.com>"]
license = "BSD-3-clause"
include = ["typed_graphql/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"
graphql-core = "3.2.1"
typing-inspect = "^0.7.1"
pytypes = "^1.0b10"
docstring-parser = "^0.14.1"

[tool.poetry.dev-dependencies]
pudb = "^2019.2"
pytest = "^6.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
